% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bits_encoding}
\alias{bits_encoding}
\title{Create a bits matrix}
\usage{
bits_encoding(x, sep = ":")
}
\arguments{
\item{x}{A character vector.}

\item{sep}{A character used to separate the group names (default is \code{:}).}
}
\value{
A numeric matrix with values of 0 or 1.
}
\description{
Produce a bits matrix for all possible combinations of the input sets.
}
\examples{
bits_encoding(c("qqa", "bnk", "sdf", "123"))
}
