% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{ellipse_line}
\alias{ellipse_line}
\alias{ellipse_fill}
\title{Control the ellipse textures}
\usage{
ellipse_line(
  linetype = "blank",
  linewidth = 0.5,
  color = c("#CC79A7", "#009E73", "#E69F00", "#56B4E9"),
  alpha = 0.5
)

ellipse_fill(
  color = c("#CC79A7", "#009E73", "#E69F00", "#56B4E9"),
  alpha = 0.2
)
}
\arguments{
\item{linetype}{A character.
The options are: blank, solid, dashed, dotted, dotdash, longdash, twodash.}

\item{linewidth}{A number (default: 0.5).}

\item{color}{A character vector (default: \code{c("#009E73", "#E69F00", "#CC79A7", "#56B4E9")}).}

\item{alpha}{A number range from 0 to 1 (default: 0.5). Set the line transparency.}
}
\value{
A list contains "color" and "alpha" values.
}
\description{
This is an internal function used to control the ellipse line color and transparency.
}
\seealso{
\code{\link[=ellipse_fill]{ellipse_fill()}}
}
