% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{ellipse_position}
\alias{ellipse_position}
\title{Ellipse positions}
\usage{
ellipse_position()
}
\value{
A two-layer nested list.
}
\description{
This is an internal function used to generate the ellipse parameters that
was used by the \code{generate_ellipse_path()} function.
It returns a list encompassing 3 lists.
Each list consists of the x0, y0, long-arm, short-arm, and angle default values
for generating 2, 3, or 4 ellipses.
}
\examples{
ellipse_position()
}
