% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ellipse.R
\name{generate_ellipse_1}
\alias{generate_ellipse_1}
\title{Generate Ellipse Path}
\usage{
generate_ellipse_1(
  x0 = 0,
  y0 = 0,
  a = 2,
  b = 1,
  angle = 0,
  density = 300,
  group = "e1"
)
}
\arguments{
\item{x0}{(default: 0) The coordinate x of the ellipse center point.}

\item{y0}{(default: 0) The coordinate y of the ellipse center point.}

\item{a}{(default: 2) Long arm.}

\item{b}{(default: 1) Short arm.}

\item{angle}{(default: 0) Rotation angle (degree).}

\item{density}{(default: 300) Higher value generate smoother ellipse.}

\item{group}{(default: "e1") Name for this ellipse.}
}
\value{
A data.frame with two columns, x and y.
The coordinates of the ellipse points to construct the ellipse.
}
\description{
Generate Ellipse Path
}
\examples{
library(ggplot2)
e1 <- generate_ellipse_1(x0 = 1, angle = 20, group = "e1")
e2 <- generate_ellipse_1(group = "e2")
p1 <- rbind(e1, e2)
ggplot(p1, aes(x, y, color = group)) +
    geom_polygon(fill = NA) +
    coord_fixed()
}
