% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ellipse.R
\name{generate_ellipse_3}
\alias{generate_ellipse_3}
\title{Plot a 3 sets venn diagram}
\usage{
generate_ellipse_3(
  data = list(`Set A` = c(1, 2, 3, 4, 5, 7, 9, 10), `Set B` = c(3, 5, 7, 9, 11, 12, 15),
    `Set C` = c(1, 4, 5, 7, 9, 10, 13)),
  x0 = c(0, -0.45, 0.45),
  y0 = c(0.45, -0.45, -0.45),
  a = c(1, 1, 1),
  b = c(1, 1, 1),
  ellipse_angle = c(0, 0, 0),
  ellipse_density = c(100, 100, 100),
  ellipse_line_color = rep("transparent", 3),
  ellipse_line_width = 0,
  ellipse_fill_color = c("#D55E00", "#009E73", "#56B4E9"),
  ellipse_fill_alpha = 0.15,
  show_set_label = TRUE,
  set_label_position = list(x = c(0, -1.4, 1.4), y = c(1.65, 0.3, 0.3)),
  set_label_family = rep("sans", times = 3),
  set_label_size = rep(5, times = 3),
  set_label_face = rep("bold", times = 3),
  set_label_color = c("#D55E00", "#009E73", "#56B4E9"),
  set_label_angle = c(0, 45, -45),
  show_zone_label = TRUE,
  zone_label = c("A", "B", "C", "AB", "AC", "BC", "ABC"),
  zone_label_position = list(x = c(A = 0, B = -1.2, C = 1.2, AB = -0.6, AC = 0.6, BC = 0,
    ABC = 0), y = c(A = 1.3, B = -0.2, C = -0.2, AB = 0.4, AC = 0.4, BC = -0.65, ABC =
    0.2)),
  zone_label_family = rep("sans", times = 7),
  zone_label_size = rep(3, times = 7),
  zone_label_face = rep("bold.italic", times = 7),
  zone_label_color = rep("black", times = 7),
  zone_label_angle = rep(0, times = 7),
  show_zone_count = TRUE,
  zone_count_position = list(x = c(A = 0, B = -0.9, C = 0.9, AB = -0.65, AC = 0.65, BC =
    0, ABC = 0), y = c(A = 1, B = -0.5, C = -0.5, AB = 0.15, AC = 0.15, BC = -0.9, ABC =
    -0.1)),
  zone_count_hjust = 0,
  zone_count_vjust = 0,
  zone_count_family = rep("sans", times = 7),
  zone_count_size = rep(4, times = 7),
  zone_count_face = rep("plain", times = 7),
  zone_count_color = rep("grey30", times = 7),
  zone_count_angle = rep(0, times = 7),
  show_zone_percentage = TRUE,
  zone_percentage_digits = 2,
  zone_percentage_position = list(x = zone_count_position[["x"]], y =
    zone_count_position[["y"]] - 0.2),
  zone_percentage_hjust = 0,
  zone_percentage_vjust = 0,
  zone_percentage_family = rep("sans", times = 7),
  zone_percentage_size = rep(3.5, times = 7),
  zone_percentage_face = rep("plain", times = 7),
  zone_percentage_color = rep("grey40", times = 7),
  zone_percentage_angle = rep(0, times = 7)
)
}
\arguments{
\item{data}{A named list with length of 3.}

\item{x0}{The x coordinates for the center points of the ellipses (default: c(-1.1, 0, 0, 1.1)).}

\item{y0}{The y coordinates for the center points of the ellipses (default: c(-.6, 0, 0, -.6)).}

\item{a}{The length of the long arm of the ellipses (default: c(2.5, 2.5, 2.5, 2.5)).}

\item{b}{The length of the short arm of the ellipses (default: c(1.35, 1.2, 1.2, 1.35)).}

\item{ellipse_angle}{Rotation angle (degree) of the ellipses (default: c(-45, -50, 50, 45)).}

\item{ellipse_density}{Smoothness of the ellipses. Higher value yield smoother shape, but compute slower (default: 100).}

\item{ellipse_line_color}{(default: rep("transparent", 4))}

\item{ellipse_line_width}{(default: 0)}

\item{ellipse_fill_color}{(default: c("#CC79A7", "#009E73", "#E69F00", "#56B4E9"))}

\item{ellipse_fill_alpha}{(default: 0.15)}

\item{show_set_label}{Logical. Show the set names? (default: TRUE)}

\item{set_label_position}{(default: list(x = c(-2.5, -1.3, 1.3, 2.5), y = c(-1.5, 2.4, 2.4, -1.5)))}

\item{set_label_family}{(default: "sans")}

\item{set_label_size}{(default: 5)}

\item{set_label_face}{(default: "bold")}

\item{set_label_color}{(default: c("#CC79A7", "#009E73", "#E69F00", "#56B4E9"))}

\item{set_label_angle}{(default: c(-45, 0, 0, 45))}

\item{show_zone_label}{Logical. Show the zone names? (default: TRUE)}

\item{zone_label}{A character vector with length of 15. The zone names.}

\item{zone_label_position}{A list contains two numeric vectors with length of 15, each for x and y coordinates. Similar with \code{set_label_position}.}

\item{zone_label_family}{(default: "sans")}

\item{zone_label_size}{(default: 3)}

\item{zone_label_face}{(default: "bold.italic")}

\item{zone_label_color}{(default: "black")}

\item{zone_label_angle}{(default: 0)}

\item{show_zone_count}{Logical. Show the number of elements for each zone? (default: TRUE)}

\item{zone_count_position}{Similar with \code{zone_label_position}.}

\item{zone_count_hjust}{Horizontally adjust the zone_count label position.}

\item{zone_count_vjust}{Vertically adjust the zone_count label position.}

\item{zone_count_family}{(default: "sans")}

\item{zone_count_size}{(default: 4)}

\item{zone_count_face}{(default: "plain")}

\item{zone_count_color}{(default: "grey30")}

\item{zone_count_angle}{(default: 0)}

\item{show_zone_percentage}{Logical. Show the percentage of \code{zone_count} for each zone? (default: TRUE).}

\item{zone_percentage_digits}{Rounding for the percentage (default: 2).}

\item{zone_percentage_position}{Similar with \code{zone_count_position}.}

\item{zone_percentage_hjust}{Similar with \code{zone_count_hjust}.}

\item{zone_percentage_vjust}{Similar with \code{zone_count_vjust}.}

\item{zone_percentage_family}{(default: "sans")}

\item{zone_percentage_size}{(default: 3.5)}

\item{zone_percentage_face}{(default: "plain")}

\item{zone_percentage_color}{(default: "grey50")}

\item{zone_percentage_angle}{(default: 0)}
}
\value{
A list which contains:
\enumerate{
\item plot: A venn diagram;
\item ellipse_path: Polygon path of each ellipses;
\item venny_df: A summary table;
\item set_elements: Elements for each sets;
\item zone_elements: Elements for each subsets.
}
}
\description{
Generate a list which contains:
\enumerate{
\item Venn plot;
\item Polygon path of each ellipses;
\item a summary table
\item Elements for each sets
\item Elements for each subsets
}
}
\examples{
s <- 1:1000
venn_list <- list(
    Set_A = sample(s, 500),
    Set_B = sample(s, 800),
    Set_C = sample(s, 700)
)
out <- generate_ellipse_3(venn_list)
out$plot
}
