% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ellipse.R
\name{generate_ellipse_4}
\alias{generate_ellipse_4}
\title{Plot a venn diagram}
\usage{
generate_ellipse_4(
  data = list(`Set A` = c(), `Set B` = c(), `Set C` = c(), `Set D` = c()),
  x0 = c(-1.1, 0, 0, 1.1),
  y0 = c(-0.6, 0, 0, -0.6),
  a = c(2.5, 2.5, 2.5, 2.5),
  b = c(1.35, 1.2, 1.2, 1.35),
  ellipse_angle = c(-45, -50, 50, 45),
  ellipse_density = c(100, 100, 100, 100),
  ellipse_line_color = rep("transparent", 4),
  ellipse_line_width = 0,
  ellipse_fill_color = c("#CC79A7", "#009E73", "#E69F00", "#56B4E9"),
  ellipse_fill_alpha = 0.15,
  set_label = TRUE,
  set_label_position = list(x = c(-2.5, -1.3, 1.3, 2.5), y = c(-1.5, 2.4, 2.4, -1.5)),
  set_label_family = rep("sans", times = 4),
  set_label_size = rep(5, times = 4),
  set_label_face = rep("bold", times = 4),
  set_label_color = c("#CC79A7", "#009E73", "#E69F00", "#56B4E9"),
  set_label_angle = c(-45, 0, 0, 45),
  zone_label = TRUE,
  zone_label_position = list(x = c(A = -2.4, B = -1.3, C = 1.3, D = 2.4, AB = -1.6, BC =
    0, CD = 1.6, AC = -1.4, AD = 0, BD = 1.4, ABC = -0.7, BCD = 0.7, ACD = -0.7, ABD =
    0.7, ABCD = 0), y = c(A = 1.15, B = 1.9, C = 1.9, D = 1.15, AB = 1.15, BC = 1.4, CD =
    1.15, AC = -0.5, AD = -1.8, BD = -0.5, ABC = 0.5, BCD = 0.5, ACD = -1.1, ABD = -1.1,
    ABCD = -0.3)),
  zone_label_family = rep("sans", times = 15),
  zone_label_size = rep(3, times = 15),
  zone_label_face = rep("bold.italic", times = 15),
  zone_label_color = rep("black", times = 15),
  zone_label_angle = rep(0, times = 15),
  zone_count = TRUE,
  zone_count_position = list(x = c(A = -2.4, B = -0.9, C = 0.9, D = 2.4, AB = -1.5, BC =
    0, CD = 1.5, AC = -1.4, AD = 0, BD = 1.4, ABC = -0.7, BCD = 0.7, ACD = -0.7, ABD =
    0.7, ABCD = 0), y = c(A = 0.5, B = 1.7, C = 1.7, D = 0.5, AB = 0.8, BC = 1.1, CD =
    0.8, AC = -0.8, AD = -2.1, BD = -0.8, ABC = 0.2, BCD = 0.2, ACD = -1.4, ABD = -1.4,
    ABCD = -0.6)),
  zone_count_hjust = 0,
  zone_count_vjust = 0,
  zone_count_family = rep("sans", times = 15),
  zone_count_size = rep(4, times = 15),
  zone_count_face = rep("plain", times = 15),
  zone_count_color = rep("grey30", times = 15),
  zone_count_angle = rep(0, times = 15),
  zone_percentage = TRUE,
  zone_percentage_digits = 2,
  zone_percentage_position = list(x = zone_count_position[["x"]], y =
    zone_count_position[["y"]] - 0.3),
  zone_percentage_hjust = 0,
  zone_percentage_vjust = 0,
  zone_percentage_family = rep("sans", times = 15),
  zone_percentage_size = rep(3.5, times = 15),
  zone_percentage_face = rep("plain", times = 15),
  zone_percentage_color = rep("grey50", times = 15),
  zone_percentage_angle = rep(0, times = 15)
)
}
\arguments{
\item{data}{A named list.}

\item{x0}{The x coordinates for the center points of the ellipses (default: c(-1.1, 0, 0, 1.1)).}

\item{y0}{The y coordinates for the center points of the ellipses (default: c(-.6, 0, 0, -.6)).}

\item{a}{The length of the long arm of the ellipses (default: c(2.5, 2.5, 2.5, 2.5)).}

\item{b}{The length of the short arm of the ellipses (default: c(1.35, 1.2, 1.2, 1.35)).}

\item{ellipse_angle}{Rotation angle (degree) of the ellipses (default: c(-45, -50, 50, 45)).}

\item{ellipse_density}{Smoothness of the ellipses. Higher value yield smoother shape, but compute slower (default: 100).}

\item{ellipse_line_color}{(default: rep("transparent", 4))}

\item{ellipse_line_width}{(default: 0)}

\item{ellipse_fill_color}{(default: c("#CC79A7", "#009E73", "#E69F00", "#56B4E9"))}

\item{ellipse_fill_alpha}{(default: 0.15)}

\item{set_label}{Show the names of each sets (default: TRUE).}

\item{set_label_position}{(default: list(x = c(-2.5, -1.3, 1.3, 2.5), y = c(-1.5, 2.4, 2.4, -1.5)))}

\item{set_label_family}{(default: "sans")}

\item{set_label_size}{(default: 5)}

\item{set_label_face}{(default: "bold")}

\item{set_label_color}{(default: c("#CC79A7", "#009E73", "#E69F00", "#56B4E9"))}

\item{set_label_angle}{(default: c(-45, 0, 0, 45))}

\item{zone_label}{Show the zone names (default: TRUE).}

\item{zone_label_position}{Similar with \code{set_label_position.} A named list consists of \code{x} and \code{y} coordinates for each zone labels.}

\item{zone_label_family}{(default: "sans")}

\item{zone_label_size}{(default: 3)}

\item{zone_label_face}{(default: "bold.italic")}

\item{zone_label_color}{(default: "black")}

\item{zone_label_angle}{(default: 0)}

\item{zone_count}{Show the number of elements for each zone? (default: TRUE)}

\item{zone_count_position}{Similar with \code{zone_label_position}.}

\item{zone_count_hjust}{Horizontally adjust the zone_count label position.}

\item{zone_count_vjust}{Vertically adjust the zone_count label position.}

\item{zone_count_family}{(default: "sans")}

\item{zone_count_size}{(default: 4)}

\item{zone_count_face}{(default: "plain")}

\item{zone_count_color}{(default: "grey30")}

\item{zone_count_angle}{(default: 0)}

\item{zone_percentage}{Show the percentage of \code{zone_count} for each zone? (default: TRUE).}

\item{zone_percentage_digits}{Rounding for the percentage (default: 2).}

\item{zone_percentage_position}{Similar with \code{zone_count_position}.}

\item{zone_percentage_hjust}{Similar with \code{zone_count_hjust}.}

\item{zone_percentage_vjust}{Similar with \code{zone_count_vjust}.}

\item{zone_percentage_family}{(default: "sans")}

\item{zone_percentage_size}{(default: 3.5)}

\item{zone_percentage_face}{(default: "plain")}

\item{zone_percentage_color}{(default: "grey50")}

\item{zone_percentage_angle}{(default: 0)}
}
\value{
A list which contains:
\enumerate{
\item plot: A venn diagram;
\item ellipse_path: Polygon path of each ellipses;
\item venny_df: A summary table;
\item set_elements: Elements for each sets;
\item zone_elements: Elements for each subsets.
}
}
\description{
Generate a list which contains:
\enumerate{
\item Venn plot;
\item Polygon path of each ellipses;
\item a summary table
\item Elements for each sets
\item Elements for each subsets
}
}
\examples{
library(ggplot2)
s <- 1:1000
venn_list <- list(
    Set_A = sample(s, 500),
    Set_B = sample(s, 800),
    Set_C = sample(s, 700),
    Set_D = sample(s, 600)
)
out <- generate_ellipse_4(venn_list)
out$plot + labs(title = "Four sets venn diagram")
}
