% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{how_many_subsets}
\alias{how_many_subsets}
\title{How many subsets}
\usage{
how_many_subsets(x, detail = FALSE)
}
\arguments{
\item{x}{A character vector.}

\item{detail}{Logical (default: FALSE). Whether to output the combinations.}
}
\value{
An integer value or a list.
}
\description{
Calculate the number of combinations that can be derived from the given character vector.
}
\examples{
how_many_subsets(c("qqa", "bnk", "sdf", "123"))
}
