% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{iter_setops}
\alias{iter_setops}
\alias{intersect_n}
\alias{union_n}
\alias{setdiff_n}
\title{Iterative set operation}
\usage{
intersect_n(lst)

union_n(lst)

setdiff_n(lst)
}
\arguments{
\item{lst}{A list with at least length of 2.}
}
\value{
An atomic vector.
}
\description{
Proceed set operation along a list.
}
\examples{
intersect_n(list(1:3, 3:5, 7:9))
union_n(list(1:3, 3:5, 7:9))
setdiff_n(list(1:3, 3:5, 7:9))
}
