% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LGL23}
\alias{LGL23}
\title{RNA-seq data}
\format{
An object of class \code{list} of length 4.
}
\usage{
LGL23
}
\description{
A list of essential information for RNA-seq analysis.
}
\details{
\itemize{
\item count_matrix: A matrix of the gene expression count reads for each sample.
The row names are gene IDs; the column names are sample IDs.
\item sample_info: the information for the sample IDs noted in the count matrix.
\item GFD: A data.frame of gene functional description.
}
}
\keyword{datasets}
