% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{set_label_default}
\alias{set_label_default}
\alias{subset_label_default}
\title{Set labels}
\usage{
set_label_default(n_sets)

subset_label_default(n_sets)
}
\arguments{
\item{n_sets}{An integer.}
}
\value{
A list.
}
\description{
This is an internal function for \code{venny()} to automatically generate
set labels, when the input list is unnamed.
}
\examples{
set_label_default(3)
}
