% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{set_label_font}
\alias{set_label_font}
\alias{subset_label_font}
\alias{subset_count_font}
\alias{subset_percentage_font}
\title{Set label font}
\usage{
set_label_font(
  family = "sans",
  face = "bold",
  size = 5,
  color = c("#CC79A7", "#009E73", "#E69F00", "#56B4E9"),
  angle = NULL
)

subset_label_font(
  family = "sans",
  face = "bold.italic",
  size = 4,
  color = "black",
  angle = 0
)

subset_count_font(
  family = "sans",
  face = "plain",
  size = 4,
  color = "grey20",
  angle = 0
)

subset_percentage_font(
  family = "sans",
  face = "plain",
  size = 4,
  color = "grey20",
  angle = 0
)
}
\arguments{
\item{family}{Character (default: "sans").}

\item{face}{Character (default: "bold").}

\item{size}{Numeric (default: 5).}

\item{color}{Character (default: c("#009E73", "#E69F00", "#CC79A7", "#56B4E9")).}

\item{angle}{Numeric | NULL (default: NULL).}
}
\value{
A list.
}
\description{
Generate a list of the available set label font parameters.
}
\examples{
set_label_font()
}
