% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{set_label_position}
\alias{set_label_position}
\alias{subset_label_position}
\alias{subset_count_position}
\title{Set label position}
\usage{
set_label_position(hjust = 0, vjust = 0, show = NULL, hide = NULL)

subset_label_position(hjust = 0, vjust = 0, show = NULL, hide = NULL)

subset_count_position(hjust = 0, vjust = 0, show = NULL, hide = NULL)
}
\arguments{
\item{hjust}{A numeric vector (default: 0).
Horizontal adjustment, adjust the x-axis coordinates of the set labels.}

\item{vjust}{A numeric vector (default: 0).
Vertical adjustment, adjust the y-axis coordinates of the set labels.}

\item{show}{A character vector (default: NULL).
Show only the specified set labels. By default, show all labels.}

\item{hide}{A character vector (default: NULL).
Do not show the specified set labels. By default, show all labels.}
}
\value{
A list contains 3 named list.
Each named list contains the x, y coordinates of the set labels for
different venn diagram (2, 3, or 4 ellipses), respectively.
}
\description{
Generate the coordinates for the set labels.
}
\examples{
set_label_position()
}
