% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{union_n}
\alias{union_n}
\alias{intersect_n}
\alias{setdiff_n}
\title{Set Operation}
\usage{
union_n(lst)

intersect_n(lst)

setdiff_n(lst)
}
\arguments{
\item{lst}{A list with at least length of 2.}
}
\value{
An atomic vector
}
\description{
Proceed iterative set operation along a list.
}
\examples{
union_n(list(1:3, 3:5, 7:9))
}
\seealso{
\link{intersect_n}, \link{setdiff_n}
}
