% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{venn_summary}
\alias{venn_summary}
\title{Create a \code{venny} summary data.frame}
\usage{
venn_summary(data = list(), show_elements = TRUE)
}
\arguments{
\item{data}{A named list contains 2 ~ 26 sets of elements.
For example, \code{list(set1 = 1:5, set2 = 3:7)}.}

\item{show_elements}{Logical (default: FALSE).
List out the elements of each zone in the dataframe.}
}
\value{
A data frame contains necessary information for subsequent inferences and plotting.
}
\description{
A data.frame consists of each combination of the input sets.
}
\examples{
s <- 1:150
venn_list <- list(
    sample(s, 50),
    sample(s, 37),
    sample(s, 46),
    sample(s, 41)
)
df0 <- venn_summary(venn_list)
df0
}
