% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{venny_df}
\alias{venny_df}
\title{Create a \code{venny} data.frame}
\usage{
venny_df(sets = list())
}
\arguments{
\item{sets}{A list contains 2 ~ 26 sets of elements.
For example, \code{list(set1 = 1:5, set2 = 3:7)}.}
}
\value{
A data frame contains necessary information for subsequent inferences and plotting.
}
\description{
A data.frame consists of each combination of the input sets.
}
\examples{
s <- 1:1000
venn_list <- list(
    Set_A = sample(s, 500),
    Set_B = sample(s, 800),
    Set_C = sample(s, 700),
    Set_D = sample(s, 600)
)
df0 <- venny_df(venn_list)
df0
}
