% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{venny_init}
\alias{venny_init}
\title{Initialize an \code{venny} object}
\usage{
venny_init(sets = list())
}
\arguments{
\item{sets}{A list contains 2 ~ 26 sets of elements.
For example, \code{list(set1 = 1:5, set2 = 3:7)}.}
}
\value{
A data frame contains necessary information for subsequent inferences and plotting.
}
\description{
Initialize an \code{venny} object
}
\examples{
lst_4sets <- list(
set_1 = letters[1:26],
set_2 = letters[3:17],
set_3 = letters[11:24],
set_4 = letters[1:20]
)
df0 <- venny_init(lst_4sets)
head(df0)
#>    A B C D length percentage elements
#> A  1 0 0 0      2   7.692308      y,z
#> B  0 1 0 0      0   0.000000
#> AB 1 1 0 0      0   0.000000
#> C  0 0 1 0      0   0.000000
#> AC 1 0 1 0      4  15.384615  u,v,w,x
#> BC 0 1 1 0      0   0.000000
}
