% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bits_encoding}
\alias{bits_encoding}
\alias{calc_zone_n}
\title{Bits matrix}
\usage{
bits_encoding(set_names)

calc_zone_n(set_names)
}
\arguments{
\item{set_names}{Character vector}
}
\value{
A numeric matrix with values of 0 or 1.
}
\description{
Produce a bits matrix for all possible combinations of the input sets.
}
\examples{
bits_encoding(c("A", "B", "C", "D"))
}
