% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LGL23}
\alias{LGL23}
\title{RNA-seq data}
\format{
\subsection{\code{LGL23}}{

A matrix and a data.frame.
For the matrix -- \code{cmat}:
\describe{
rownames: gene ID
colnames: sample ID
}
}
}
\usage{
LGL23
}
\description{
A matrix (read count data) and a data.frame (information for the samples,
including \code{genotype}, \code{treatment}, and \code{group} columns.)
}
\keyword{datasets}
