% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_related.R
\name{polygon_union}
\alias{polygon_union}
\alias{polygon_intersect}
\alias{polygon_setdiff}
\title{Polygon manipulation}
\usage{
polygon_union(poly1, poly2)

polygon_intersect(poly1, poly2)

polygon_setdiff(poly1, poly2)
}
\arguments{
\item{poly1}{The polygon acts as the subject. The input should be a matrix.
The first column is the x coordinates, while
the second column is the y coordinates.}

\item{poly2}{The polygon which performs the manipulation.
The first column is the x coordinates, while
the second column is the y coordinates.}
}
\value{
A matrix.
}
\description{
Perform union, intersection, and difference between two polygons.
}
\examples{
library(ggplot2)
s <- 1:1000
venn_list <- list(
    Set_A = sample(s, 500),
    Set_B = sample(s, 800),
    Set_C = sample(s, 700),
    Set_D = sample(s, 600)
)
venn <- generate_ellipse_4(venn_list, ellipse_fill_alpha = 0.1)
e4 <- venn$ellipse_path
poly <- polygon_union(e4$`Set_A`, e4$`Set_C`) |>
    polygon_intersect(e4$`Set_B`) |>
    polygon_setdiff(e4$`Set_D`)
# Add a polygon to the venn plot
venn$plot +
    geom_polygon(
        inherit.aes = FALSE,
        data = poly,
        mapping = aes(x, y),
        fill = "red",
        alpha = 0.4,
        color = "maroon",
        linewidth = 1
    )
}
