% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venny.R
\name{venny}
\alias{venny}
\title{Venn diagram}
\usage{
venny(
  data,
  detail = FALSE,
  ellipse.line = ellipse_line(),
  ellipse.fill = ellipse_fill(),
  ellipse.density = 200,
  set.label = TRUE,
  set.label.position = set_label_position(),
  set.label.font = set_label_font(),
  subset.label = TRUE,
  subset.label.position = subset_label_position(),
  subset.label.font = subset_label_font(),
  subset.count = TRUE,
  subset.count.position = subset_count_position(),
  subset.count.font = subset_count_font(),
  subset.percentage = TRUE,
  subset.percentage.position = subset_count_position(vjust = -0.3),
  subset.percentage.font = subset_percentage_font()
)
}
\arguments{
\item{data}{A list with 2 to 4 vectors}

\item{detail}{Logical (default: TRUE).
If TRUE, output a list contains the venn diagram and summary table.
Otherwise, output only venn diagram.}

\item{ellipse.line}{A list. See \code{ellipse_line()}.}

\item{ellipse.fill}{A list. See \code{ellipse_fill()}.}

\item{ellipse.density}{Integer (default: 200). Higher value yield smoother ellipse.}

\item{set.label}{Logical (default: TRUE). If TRUE, show the set labels.}

\item{set.label.position}{A list. See \code{set_label_position()}.}

\item{set.label.font}{A list. See \code{set_label_font()}.}

\item{subset.label}{Logical (default: TRUE). If TRUE, show the subset labels.
If a named list is provided, then the selected subset name will be renamed.
For example, list(AB = "new_AB") will change the original subset AB name to "new_AB".}

\item{subset.label.position}{A list. See \code{subset_label_position()}.}

\item{subset.label.font}{A list. See \code{subset_label_font()}.}

\item{subset.count}{Logical (default: TRUE). If TRUE, show the element counts for each subset.}

\item{subset.count.position}{A list. See \code{subset_count_position()}.}

\item{subset.count.font}{A list. See \code{subset_count_font()}.}

\item{subset.percentage}{Logical (default: TRUE). If TRUE, show the percentages of the counts.}

\item{subset.percentage.position}{A list. See \code{subset_percentage_position()}.}

\item{subset.percentage.font}{A list. See \code{subset_percentage_font()}.}
}
\value{
A venn diagram which is a \code{ggplot} object.
}
\description{
Venn diagram
}
\examples{
lst <- LGL23$DEGs
venny(lst)
}
