% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venny-api.R
\name{venny}
\alias{venny}
\title{Venn Diagram}
\usage{
venny(data, ...)
}
\arguments{
\item{data}{A named list consists of at least two sets of elements.}

\item{...}{Arguments pass to generate_ellipse_*() functions.}
}
\value{
A list which contains:
\enumerate{
\item plot: A venn diagram;
\item ellipse_path: Polygon path of each ellipses;
\item venny_df: A summary table;
\item set_elements: Elements for each sets;
\item zone_elements: Elements for each subsets.
}
}
\description{
A wrapper for generate_ellipse_*() functions.
}
\examples{
s <- 1:1000
lst <- list(
    Set_A = sample(s, 500),
    Set_B = sample(s, 700),
    Set_C = sample(s, 900),
    Set_D = sample(s, 600)
)
e3 <- venny(lst[1:3])
e3$plot
e4 <- venny(lst[1:4], show_zone_percentage = FALSE)
e4$plot
}
\seealso{
\link{generate_ellipse_2}, \link{generate_ellipse_3}, \link{generate_ellipse_4}
}
